+++ include/common.hpp
@@ -38,9 +38,9 @@ typedef struct proposal_and_edge_counts_t {
 } ProposalAndEdgeCounts;
 
 /// Calculates the entropy of a single blockmodel cell.
-inline double cell_entropy(double value, double degree_in, double degree_out) {
+inline float cell_entropy(float value, float degree_in, float degree_out) {
     if (value == 0.0) return 0.0;
-    double entropy = value * std::log(value / (degree_in * degree_out));
+    float entropy = value * logf(value / (degree_in * degree_out));
     // if (std::isnan(entropy) || std::isinf(entropy)) {
     //     std::cerr << "value: " << value << " dIn: " << degree_in << " dOut: " << degree_out << std::endl;
     //     throw std::invalid_argument("something is wrong");
